ref(test): Trigger top-level errors inside sandbox.
