--- dev-packages/browser-integration-tests/suites/public-api/instrumentation/onError/non-string-arg/subject.js
@@ -1,8 +1,0 @@
-function run() {
-  window.onerror({
-    type: 'error',
-    otherKey: 'hi',
-  });
-}
-
-run();
