+++ dev-packages/browser-integration-tests/suites/public-api/instrumentation/onError/non-string-arg/test.ts
@@ -2,14 +2,25 @@ import { expect } from '@playwright/test';
 import type { Event } from '@sentry/types';
 
 import { sentryTest } from '../../../../../utils/fixtures';
-import { getFirstSentryEnvelopeRequest } from '../../../../../utils/helpers';
+import { getFirstSentryEnvelopeRequest, runScriptInSandbox } from '../../../../../utils/helpers';
 
 sentryTest(
   'should catch onerror calls with non-string first argument gracefully',
   async ({ getLocalTestPath, page }) => {
     const url = await getLocalTestPath({ testDir: __dirname });
 
-    const eventData = await getFirstSentryEnvelopeRequest<Event>(page, url);
+    await page.goto(url);
+
+    runScriptInSandbox(page, {
+      content: `
+        throw {
+          type: 'Error',
+          otherKey: 'otherValue',
+        };
+      `,
+    });
+
+    const eventData = await getFirstSentryEnvelopeRequest<Event>(page);
 
     expect(eventData.exception?.values).toHaveLength(1);
     expect(eventData.exception?.values?.[0]).toMatchObject({
