--- dev-packages/browser-integration-tests/suites/public-api/instrumentation/onError/syntax-errors/subject.js
@@ -1,10 +1,0 @@
-function run() {
-  try {
-    eval('foo{};');
-  } catch (e) {
-    // simulate window.onerror without generating a Script error
-    window.onerror('error', 'file.js', 1, 1, e);
-  }
-}
-
-run();
