--- dev-packages/browser-integration-tests/suites/public-api/instrumentation/onError/thrown-errors/subject.js
@@ -1,10 +1,0 @@
-function run() {
-  try {
-    throw new Error('realError');
-  } catch (e) {
-    // simulate window.onerror without generating a Script error
-    window.onerror('error', 'file.js', 1, 1, e);
-  }
-}
-
-run();
