--- dev-packages/browser-integration-tests/suites/public-api/startSpan/error-sync/subject.js
@@ -1,9 +1,0 @@
-function run() {
-  Sentry.startSpan({ name: 'parent_span' }, () => {
-    throw new Error('Sync Error');
-  });
-}
-
-// using `setTimeout` here because otherwise the thrown error will be
-// thrown as a generic "Script Error." instead of the actual error".
-setTimeout(run);
