--- dev-packages/browser-integration-tests/suites/tracing/browserTracingIntegration/error/subject.js
@@ -1,3 +1,0 @@
-setTimeout(() => {
-  throw new Error('Error during pageload');
-}, 100);
